requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.7
django-extensions==3.2.3
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
python-dotenv